Fix for CenterWithin not working as expected
The previous implementation of CenterWithin didn't leave the cursor in the expected position. By creting a Dummy it would reserve the required space and advance the cursor the correct distance as a normal widget would, but we would then move the cursor to the start of the content drawing and not set it back to the final position. The flow should be Dummy -> move cursor for content -> draw content -> move cursor to the end of the container.

As the drawing of the content needs to happen before the cursor is set to the end position it required an API change so we can render the content before setting the final cursor position. An alternate implementation would be to return a delegate that needs to be called after the content is drawn so the cursor is advanced, but that just seems much more complex and requires more of the user than is required.

Because we need to advance the cursor to an end point it didn't make sense to continue supporting the overloads which took only a width. Previously we would create a dummy with 0 height which would not play well with advancing the cursor. The existing functionality in the demo has been preserved with minor tweaks, such as getting the label size and not just the width.

This implementation has correct spacing and works with items rendered via CenterWithin on consequtive lines, and works if they are on the same line (ImGui.SameLine)