Reverted recent changes to why the Dummy widget is rendered
By rendering the Dummy last we get perfect layout logic, but the Dummy is rendered last and is considered first for things like clicks and hover events. Rendering the Dummy first means usages of ImGui.SameLine won't work, but the custom content is rendered on top and will behave correctly. Updated the demo to remove the broken SameLine case.